+++ src/common/component/DatePicker/Calendar/Days/CalendarDays.tsx
@@ -1,4 +1,4 @@
-import { datesContainerStyle, dayStyle } from '@/common/component/DatePicker/Calender/index/Calendar.style';
+import { datesContainerStyle, dayStyle } from '@/common/component/DatePicker/Calendar/Calendar.style';
 import Text from '@/common/component/Text/Text';
 
 interface CalendarDaysProps {
@@ -8,7 +8,7 @@ interface CalendarDaysProps {
 const CalendarDays = ({ weekDays }: CalendarDaysProps) => (
   <div css={datesContainerStyle}>
     {weekDays.map((day, index) => (
-      <Text tag="body8" key={index} css={dayStyle}>
+      <Text tag="body8" key={index} css={dayStyle} tabIndex={0}>
         {day}
       </Text>
     ))}
