+++ src/common/component/DatePicker/DatePickerContainer/DatePickerContainer.tsx
@@ -1,12 +1,8 @@
-import { ReactNode, forwardRef } from 'react';
+import { ComponentPropsWithRef, forwardRef } from 'react';
 
 import { containerStyle } from '@/common/component/DatePicker/DatePickerContainer/DatePickerContainer.style';
 
-interface DatePickerContainerProps {
-  children: ReactNode;
-}
-
-const DatePickerContainer = forwardRef<HTMLDivElement, DatePickerContainerProps>(({ children }, ref) => (
+const DatePickerContainer = forwardRef<HTMLDivElement, ComponentPropsWithRef<'div'>>(({ children }, ref) => (
   <div ref={ref} css={containerStyle}>
     {children}
   </div>
