+++ src/common/component/DatePicker/index.tsx
@@ -1,4 +1,4 @@
-import DatePickerCalendar from '@/common/component/DatePicker/Calender/index/DatePickerCalendar';
+import DatePickerCalendar from '@/common/component/DatePicker/Calendar/DatePickerCalendar';
 import DatePickerContainer from '@/common/component/DatePicker/DatePickerContainer/DatePickerContainer';
 import DatePickerTrigger from '@/common/component/DatePicker/Trigger/DatePickerTrigger';
